Added "silent" option to the env
